requirements.txt
annotated-types==0.7.0
anyio==4.9.0
autocommand==2.2.2
certifi==2025.1.31
cheroot==10.0.1
CherryPy==18.10.0
distro==1.9.0
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jiter==0.9.0
more-itertools==10.6.0
openai==1.71.0
portend==3.2.0
PyAudio==0.2.14
pydantic==2.11.2
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
tempora==5.8.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
websocket-client==1.8.0
ws4py==0.6.0
zc.lockfile==3.0.post1
zope.event==5.0
zope.interface==7.2
